backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-Migrate==4.0.4
Werkzeug==2.2.3
Flask-Login==0.6.2
python-dotenv==1.0.1
sqlalchemy-serializer==1.4.1